Starting attempts to clean up/refactor, starting with xkeys/extension
