--- .eslintignore
@@ -1,2 +1,0 @@
-types/**/*
-schemas/**/*
