--- .eslintrc.browser.js
@@ -1,57 +1,0 @@
-module.exports = {
-  root: true,
-  parser: 'vue-eslint-parser',
-  parserOptions: {
-    parser: '@typescript-eslint/parser',
-    project: 'tsconfig.json',
-    sourceType: 'module',
-    extraFileExtensions: ['.vue'],
-  },
-  env: {
-    es6: true,
-    node: true,
-  },
-  globals: {
-    nodecg: 'readonly',
-    NodeCG: 'readonly',
-  },
-  plugins: [
-    'vue',
-    '@typescript-eslint',
-  ],
-  extends: [
-    'airbnb-base',
-    'plugin:vue/recommended',
-    'eslint:recommended',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:import/typescript',
-  ],
-  settings: {
-    'import/core-modules': ['nodecg/types/browser'],
-    'import/extensions': ['.js', '.jsx', '.ts', '.tsx'],
-    'import/resolver': {
-      node: {
-        moduleDirectory: [
-          'node_modules',
-          '../..',
-          '..',
-          '.',
-        ],
-      },
-    },
-  },
-  rules: {
-    'import/extensions': ['error', 'ignorePackages', {
-      js: 'never',
-      jsx: 'never',
-      ts: 'never',
-      tsx: 'never',
-    }],
-    'import/no-extraneous-dependencies': ['error', {
-      packageDir: ['.', '../..'],
-    }],
-    'lines-between-class-members': 'off',
-    'max-len': ['error', { 'code': 100 }],
-    'class-methods-use-this': 'off',
-  },
-};
