--- jsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "*": [
-        "./*"
-      ]
-    }
-  },
-  "exclude": [
-    "node_modules"
-  ]
-}
