+++ package.json
@@ -11,39 +11,5 @@
     "url": "git+https://github.com/esamarathon/esa-layouts-shared.git"
   },
   "license": "MIT",
-  "author": "zoton2",
-  "scripts": {
-    "typeschemas": "node script/typeschemas.js"
-  },
-  "dependencies": {
-    "amqp-connection-manager": "^3.2.0",
-    "amqplib": "^0.5.6",
-    "clone": "^2.1.2",
-    "obs-websocket-js": "4.0.1",
-    "osc": "^2.4.1",
-    "string-similarity": "^4.0.1",
-    "uuid": "^8.1.0",
-    "ws": "^7.3.1",
-    "xkeys": "^1.1.1"
-  },
-  "devDependencies": {
-    "@esamarathon/mq-events": "github:esamarathon/mq-events",
-    "@types/amqp-connection-manager": "^2.0.10",
-    "@types/amqplib": "^0.5.13",
-    "@types/clone": "^0.1.30",
-    "@types/needle": "^2.0.4",
-    "@types/string-similarity": "^3.0.0",
-    "@types/uuid": "^8.0.0",
-    "@types/ws": "^7.2.6",
-    "@typescript-eslint/eslint-plugin": "^3.2.0",
-    "@typescript-eslint/parser": "^3.2.0",
-    "eslint": "^7.2.0",
-    "eslint-config-airbnb-base": "^14.2.0",
-    "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "json-schema-to-typescript": "^9.1.0",
-    "nodecg": "^1.7.4",
-    "typescript": "^3.9.5",
-    "vue": "^2.6.12"
-  }
+  "author": "zoton2"
 }
