--- script/typeschemas.js
@@ -1,30 +1,0 @@
-const { compileFromFile } = require('json-schema-to-typescript');
-const fs = require('fs');
-
-try {
-  const files = fs.readdirSync('./schemas');
-  const filesProcessed = [];
-
-  files.forEach((file) => {
-    if (file.match('.json$', 'i')) {
-      const fileName = file.replace(/.json$/, '');
-      console.log(fileName);
-      filesProcessed.push(fileName);
-      compileFromFile(
-        `./schemas/${fileName}.json`,
-        {
-          cwd: './schemas',
-        },
-      ).then(ts => fs.writeFileSync(`./schemas/${fileName}.d.ts`, ts));
-    }
-  });
-
-  const index = [];
-  filesProcessed.forEach((file) => {
-    index.push(`export * from './${file}';`);
-  });
-
-  fs.writeFileSync('./schemas/index.d.ts', `${index.join('\n')}\n`);
-} catch (err) {
-  console.log('no schemas');
-}
