--- types/XKeysLib.d.ts
@@ -1,5 +1,0 @@
-export namespace XKeysLib {
-  interface Config {
-    enable: boolean;
-  }
-}
