+++ types/index.d.ts
@@ -5,4 +5,3 @@ export * from './RabbitMQ';
 export * from './Restream';
 export * from './Tracker';
 export * from './X32';
-export * from './XKeysLib';
